Move button to its own folder
